Introduce tdh2easy types for tdh2 artifacts. (#5)
The current interface may be confusing for developers who are forced to interact with both tdh2easy and tdh2 artifacts. This PR introduces tdh2easy types for shares and keys to unify the interface. It will also allow for more flexible modifications to tdh2easy and (potentially) making tdh2 an internal package.